+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # VERSION: DEV
 
 * Properly added a changelog file
+* Added config option for the upgrades to change how much they do and how often they do it, closes #204